refactor: internal structured rule config
Introducing a new rule configuration format. It contains of a single JS objct definiing all aspects of the linting.

Original array configuration made sense when capabilities of the plugin were rather limited - or rather - opinionated. As consumers and contributors were adding more features, the quality of the configuration using the format started deteriorating. Parsing the array is quite annoying and opens oprotunities for bugs such as the one fixed by #2.

The new format would not be documented at the moment and advertized to consumers until the code reaches a certain threshold of maturity. In the mean time it would be used as the internal format of the parsed configuration and tests would be gradually migrated to it.

The changes add st one new test for the new config to guaratee 100% coverage. To support the test case I had to add the new config's schema to the overal rule config.